ZapIt 1.0
My first program I created to Install, Uninstall and Zap the software
out of the registry.  Zapping out the of the registry will not remove
installation directory or files.  The only leverage this program gives
is to remove software that can not be removed so you can try to
reinstall it.  This uses Microsofts MSIZap.exe